refactor(mcp): remove dead internal exports
