+++ src/mcp/channel-server-runtime.ts
@@ -1,0 +1,68 @@
+import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+import type { OpenClawConfig } from "../config/types.openclaw.js";
+import { VERSION } from "../version.js";
+import { OpenClawChannelBridge } from "./channel-bridge.js";
+import { ClaudePermissionRequestSchema, type ClaudeChannelMode } from "./channel-shared.js";
+import { getChannelMcpCapabilities, registerChannelMcpTools } from "./channel-tools.js";
+
+async function resolveMcpConfig(config: OpenClawConfig | undefined): Promise<OpenClawConfig> {
+  if (config) {
+    return config;
+  }
+  const { getRuntimeConfig } = await import("../config/config.js");
+  return getRuntimeConfig();
+}
+
+export async function createChannelMcpRuntime(
+  opts: {
+    gatewayUrl?: string;
+    gatewayToken?: string;
+    gatewayPassword?: string;
+    config?: OpenClawConfig;
+    claudeChannelMode?: ClaudeChannelMode;
+    verbose?: boolean;
+  } = {},
+): Promise<{
+  server: McpServer;
+  bridge: OpenClawChannelBridge;
+  start: () => Promise<void>;
+  close: () => Promise<void>;
+}> {
+  const cfg = await resolveMcpConfig(opts.config);
+  const claudeChannelMode = opts.claudeChannelMode ?? "auto";
+  const capabilities = getChannelMcpCapabilities(claudeChannelMode);
+  const server = new McpServer(
+    { name: "openclaw", version: VERSION },
+    capabilities ? { capabilities } : undefined,
+  );
+  const bridge = new OpenClawChannelBridge(cfg, {
+    gatewayUrl: opts.gatewayUrl,
+    gatewayToken: opts.gatewayToken,
+    gatewayPassword: opts.gatewayPassword,
+    claudeChannelMode,
+    verbose: opts.verbose ?? false,
+  });
+  bridge.setServer(server);
+
+  server.server.setNotificationHandler(ClaudePermissionRequestSchema, async ({ params }) => {
+    await bridge.handleClaudePermissionRequest({
+      requestId: params.request_id,
+      toolName: params.tool_name,
+      description: params.description,
+      inputPreview: params.input_preview,
+    });
+  });
+  registerChannelMcpTools(server, bridge);
+
+  return {
+    server,
+    bridge,
+    start: async () => {
+      await bridge.start();
+    },
+    close: async () => {
+      await bridge.close();
+      await server.close();
+    },
+  };
+}
