+++ src/mcp/channel-server.test.ts
@@ -3,9 +3,8 @@ import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { InMemoryTransport } from "@modelcontextprotocol/sdk/inMemory.js";
 import { describe, expect, test, vi } from "vitest";
 import { z } from "zod";
-import { shouldRetryInitialMcpGatewayConnect } from "./channel-bridge.js";
 import { OpenClawChannelBridge } from "./channel-bridge.js";
-import { createOpenClawChannelMcpServer } from "./channel-server.js";
+import { createChannelMcpRuntime } from "./channel-server-runtime.js";
 import { extractAttachmentsFromMessage } from "./channel-shared.js";
 
 const ClaudeChannelNotificationSchema = z.object({
@@ -25,7 +24,7 @@ const ClaudePermissionNotificationSchema = z.object({
 });
 
 async function connectMcpWithoutGateway(params?: { claudeChannelMode?: "auto" | "on" | "off" }) {
-  const serverHarness = await createOpenClawChannelMcpServer({
+  const serverHarness = await createChannelMcpRuntime({
     claudeChannelMode: params?.claudeChannelMode ?? "auto",
     config: {} as never,
     verbose: false,
@@ -84,22 +83,7 @@ function requireFirstMockCall(mock: { mock: { calls: unknown[][] } }, label: str
   return call;
 }
 
-function gatewayRequestError(retryable: boolean): Error {
-  return Object.assign(new Error(retryable ? "gateway busy" : "auth failed"), {
-    name: "GatewayClientRequestError",
-    retryable,
-  });
-}
-
 describe("openclaw channel mcp server", () => {
-  test("keeps initial MCP gateway connection alive through transient connect errors", () => {
-    expect(
-      shouldRetryInitialMcpGatewayConnect(new Error("gateway request timeout for connect")),
-    ).toBe(true);
-    expect(shouldRetryInitialMcpGatewayConnect(gatewayRequestError(true))).toBe(true);
-    expect(shouldRetryInitialMcpGatewayConnect(gatewayRequestError(false))).toBe(false);
-  });
-
   describe("gateway-backed flows", () => {
     describe("gateway integration", () => {
       test("returns conversation and message payloads in primary MCP content", async () => {
