+++ test/scripts/check-deadcode-exports.test.ts
@@ -28,6 +28,10 @@ describe("check-deadcode-exports", () => {
     expect(knipConfig.workspaces["."].entry).toContain("src/agents/sessions/extension-sdk.ts!");
   });
 
+  it("models the spawned system-agent MCP stdio entry", () => {
+    expect(knipConfig.workspaces["."].entry).toContain("src/mcp/openclaw-tools-serve.ts!");
+  });
+
   it("parses all compact export sections and expands symbol lists", () => {
     expect(
       parseKnipCompactUnusedExports(`
